pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "enkf-ppe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "plotly>=6.5.2",
    "torch>=2.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/enkf_ppe", "paths.py"]
